+++ torchmetrics/classification/matthews_corrcoef.py
@@ -66,7 +66,6 @@ class MatthewsCorrcoef(Metric):
             will be used to perform the allgather
 
     Example:
-
         >>> from torchmetrics import MatthewsCorrcoef
         >>> target = torch.tensor([1, 1, 0, 0])
         >>> preds = torch.tensor([0, 1, 0, 0])
@@ -75,6 +74,7 @@ class MatthewsCorrcoef(Metric):
         tensor(0.5774)
 
     """
+
     def __init__(
         self,
         num_classes: int,