+++ torchmetrics/utilities/data.py
@@ -14,7 +14,7 @@
 from typing import Any, Callable, List, Mapping, Optional, Sequence, Union
 
 import torch
-from torch import Tensor
+from torch import Tensor, tensor
 
 from torchmetrics.utilities.prints import rank_zero_warn
 
@@ -242,7 +242,6 @@ def get_group_indexes(idx: Tensor) -> List[Tensor]:
         A list of integer `torch.Tensor`s
 
     Example:
-
         >>> indexes = torch.tensor([0, 0, 0, 1, 1, 1, 1])
         >>> groups = get_group_indexes(indexes)
         >>> groups
@@ -256,4 +255,4 @@ def get_group_indexes(idx: Tensor) -> List[Tensor]:
             indexes[_id] += [i]
         else:
             indexes[_id] = [i]
-    return [torch.tensor(x, dtype=torch.int64) for x in indexes.values()]
+    return [tensor(x, dtype=torch.int64) for x in indexes.values()]